part 3

+++ src/pages/settings.tsx
@@ -2,7 +2,6 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faXmark, faRotateLeft } from "@fortawesome/free-solid-svg-icons";
 import React from "react";
 
-
 function updateName(rawPeriodName: string, newName: string, setPeriodNamesDB: Function) {
 	const periodNamesDB = JSON.parse(localStorage.getItem("periodNames") || "{}");
 
@@ -27,17 +26,20 @@ function checkRemovedPeriods(period: string) {
 	return removedPeriodNames.includes(period);
 }
 
-function reset(setPeriodNamesDB: Function, setRemovedPeriodsDB: Function, createAvaliablePeriodsDB: Function, scheduleDB: any) {
-	localStorage.clear();
+function reset(setPeriodNamesDB: Function, setRemovedPeriodsDB: Function, createEventsDB: Function, scheduleDB: any, setScheduleDB: Function) {
+	const result = confirm("Are you sure that you want to reset your entire schedule?");
 
-	setPeriodNamesDB(JSON.parse(createAvaliablePeriodsDB(scheduleDB)));
-	setRemovedPeriodsDB([]);
-	localStorage.setItem("removedPeriods", JSON.stringify([]));
-	localStorage.setItem("periodNames", createAvaliablePeriodsDB(scheduleDB));
+	if (result) {
+		localStorage.clear();
 
+		setPeriodNamesDB(JSON.parse(createEventsDB(scheduleDB)));
+		setRemovedPeriodsDB([]);
+		localStorage.setItem("removedPeriods", JSON.stringify([]));
+		localStorage.setItem("periodNames", createEventsDB(scheduleDB));
+	}
 }
 
-function PeriodNameCustomizer(props: { periodName: string; rawPeriodName: string, setPeriodNamesDB: Function, setRemovedPeriodsDB: Function }) {
+function PeriodNameCustomizer(props: { periodName: string; rawPeriodName: string; setPeriodNamesDB: Function; setRemovedPeriodsDB: Function }) {
 	return (
 		<div key={props.rawPeriodName} className="flex flex-row justify-center place-items-center items-center rounded bg-wedgewood-300 lg:m-3 lg:p-5">
 			<h1 className="py-5 lg:p-6 text-3xl font-bold">{props.rawPeriodName}:</h1>
@@ -55,11 +57,11 @@ function PeriodNameCustomizer(props: { periodName: string; rawPeriodName: string
 }
 
 // @ts-ignore
-function createAvaliablePeriodsDB(scheduleDB) {
-	const avaliablePeriods = scheduleDB["about"]["avaliablePeriods"];
+function createEventsDB(scheduleDB) {
+	const allEvents = scheduleDB["about"]["allEvents"];
 	let entry: Record<string, null> = {};
 
-	for (const period of avaliablePeriods) {
+	for (const period of allEvents) {
 		entry[period] = null;
 	}
 
@@ -78,14 +80,13 @@ export default function Settings() {
 	const [removedPeriodsDB, setRemovedPeriodsDB] = React.useState(null);
 	const [isLoading, setLoading] = React.useState(true);
 
-	
 	React.useEffect(() => {
 		fetch("/api/schedule/dvhs")
 			.then((res) => res.json())
 			.then((data) => {
 				setScheduleDB(data);
 
-				if (!localStorage.getItem("periodNames")) createAvaliablePeriodsDB(data);
+				if (!localStorage.getItem("periodNames")) createEventsDB(data);
 				if (!localStorage.getItem("removedPeriods")) createRemovedPeriodsDB(setRemovedPeriodsDB);
 
 				setPeriodNamesDB(JSON.parse(localStorage.getItem("periodNames") || ""));
@@ -125,7 +126,7 @@ export default function Settings() {
 		<div className="container mx-auto lg:mt-10 flex flex-col justify-center lg:p-8 shadow-xl bg-wedgewood-200 ">
 			<div className="flex justify-between items-center">
 				<h1 className="font-bold text-3xl m-4">Customize Period Names</h1>
-				<button onClick={()=>reset(setPeriodNamesDB, setRemovedPeriodsDB, createAvaliablePeriodsDB, scheduleDB)}>
+				<button onClick={() => reset(setPeriodNamesDB, setRemovedPeriodsDB, createEventsDB, scheduleDB, setScheduleDB)}>
 					<FontAwesomeIcon icon={faRotateLeft} size="xl"></FontAwesomeIcon>
 				</button>
 			</div>
